+++ launcher/Services/LauncherVersions.cs
@@ -2,8 +2,9 @@ namespace SeapowerMultiplayer.Launcher.Services
 {
     public static class LauncherVersions
     {
-        public const string LauncherVersion = "0.6.0";
-        public const string PluginVersion = "0.1.4-alpha";
-        public const string PluginNumericVersion = "0.1.4";
+        public const string LauncherVersion = "0.6.3";
+        public const string PluginVersion = "0.1.6-alpha";
+        public const string PluginNumericVersion = "0.1.6";
+        public const int ProtocolVersion = 403;
     }
 }